Made a request to a few items, set up dicts and classes to keep track of things, and added headers to access Amazon
So amazon is good at blocking bots, so we're gonna have to follow the rules in order to maintain access. Added some header information that we'll have to randomly change in order to not get blocked again. In the meantime, we'll have to be sparing with the requests else our IP will get banned. Also added some dicts with distinct information like the selectors on the page for the necessary information as well as a class to have our pulled data be represented as an 'AmazonItem' object. From there I'll expand it out and give the class more methods like adding and organizing reviews and simply a 'save to database' kind of thing.