Random Headers and remaining attributes
Added a randomizer with a list of headers that could work to make it less likely to get blocked. Also added functions for the different selectors if they required more work to pull the data (like if one value wasn't present and so on). We could use some design pattern here to take advantage of polymorphism, but that more like a 'maybe later' kind of thing. For now it works.