Started review parsing and fixed the description element
Got the basics of the review page to start parsing, however Amazon is being difficult and preventing the scraper from 'following' to another page. So instead of figuring that out which will take too long, I'm getting the function set up first and then I'll see what I can do to get that data. Also updated the description element to pull from a better part of the Amazon page