It all works! Added a country field and pagination
I guess Amazon was a bit more forgiving at the moment and I could make multiple requests without too much trouble. Fixed a small bug with the title and added a 'country' field in case the reviews go beyond the US. The program is working as a data puller at this point.