Setting up Mongodb docker-compose and instructions for basic access
Setup a docker compose container so that a docker image could be created if a mongodb database hadn't been set up. Also wrote a test script to see if we could have python connect to the database as well as provide instructions in the notes as to how to set up the docker instance.