Basic Mongodb functionality and update changes to amazon_review_spider
Once I had a basic connection with mongodb, I added more functionality to the test to see if I could the basic CRUD operations. After that worked well, I worked on getting the pipeline operations to work with scrapy by providing the pymongo client, connecting to it, and inserting the 'quotes' test spider data. Also updated some of the mongo docker instructions for proper setup.