And it works!
Technicall, I'm done! I got it running! But I'm not really done, I need to clean things up and have it readable, but for the most part it's functioning and we're getting the data in to the database. So
what I did: 1. Configured the MongoDB information to be in the spider as opposed to just floating around (properly coupled). 2. Setup the pipeline to save the reviews into their own collection. 3. Setup the AmazonItemSpider to save a product as well as hand it's id off to the reviews. Much needs to happen in the realm of cleaning things up, but for now she works!